Major change to storage of user object. Login with Twitter.
Removed Login with Github.
Changed the logged-in user to be stored at login.user.
Changed the name of the user store to "users".
Logged-in user is stored in login and in users. Users is indexed by user id.
Updated security rules so that only the logged-in user can read secure. This
is now where credentials, social network name & pic, and email are stored.
You can only read from the users table when you know the specific
user id. This prevents an attacker from dumping a user list.